+++ Resolvers/UserUnBlockResolver.php
@@ -1,0 +1,11 @@
+<?php
+namespace App\Resolvers\User;
+use Illuminate\Auth\Events\Registered;
+
+class UserUnBlockResolver
+{
+    public static function handle($user)
+    {
+        
+    }
+}