+++ CHANGELOG.md
@@ -2,6 +2,48 @@
 
 All notable changes to `processton-user` will be documented in this file
 
-## 1.0.0 - 201X-XX-XX
+
+
+## 1.1.9 - 2024-07-12
+
+- Documentation updated
+
+## 1.1.8 - 2024-07-12
+
+- Bug Fixes (Change User Role)
+
+## 1.1.7 - 2024-07-12
+
+- Bug Fixes (Permissions update finished)
+
+## 1.1.6 - 2024-07-12
+
+- Bug Fixes (Role Permissions Edit button)
+
+## 1.1.5 - 2024-07-12
+
+- Bug Fixes (Validation issues)
+
+## 1.1.4 - 2024-07-12
+
+- Bug Fixes (Fix On Role Create)
+
+## 1.1.3 - 2024-07-12
+
+- Bug Fixes (Fix Role Edit Issue)
+
+## 1.1.2 - 2024-07-12
+
+- Bug Fixes (Fixes on role create and edit)
+
+## 1.1.1 - 2024-07-12
+
+- Bug Fixes
+
+## 1.1.0 - 2024-07-12
+
+- Roles Functionality
+
+## 1.0.0 - 2024-07-09
 
 - initial release